refactor(compress): SPL-decompose strategy/negotiator.rs
Split the 1034-line strategy/negotiator.rs into a directory module with one
submodule per concern:

- trait_def.rs: CompressionNegotiator trait
- default.rs: DefaultCompressionNegotiator (modern upstream-compatible)
- fixed.rs: FixedCompressionNegotiator (testing/--compress-choice override)
- protocol_aware.rs: ProtocolAwareCompressionNegotiator (version-gated)
- tests.rs: cross-cutting trait-object and Send/Sync tests

Tests stay co-located with each negotiator implementation. Public API
(compress::strategy::negotiator::{Compression,Default,Fixed,
ProtocolAware}CompressionNegotiator) is unchanged.

No submodule exceeds 650 LoC; the largest is default.rs at 443 lines.
